Added total, quantity, added Mappers and Welcome message.
Added a total balance of user to CustomerPage.jsp so that it's shown.
Added a quanitity to Products.jsp
Added to the CupCakeMapper: getNameOfUser, createCompletedOrder, getLatestOrderNumber(), createLineItems(Not done fully).
Added Welcome message to CustomerPage.jsp if you login or register so that it shows the name of the user.